requirements.txt
# requirements.txt

# Flask framework for building web applications
Flask==2.1.0

# Requests library for making HTTP requests
requests==2.26.0

# MySQL Connector/Python for connecting to MySQL databases
mysql-connector-python==8.0.28

# Flask-Login for managing user sessions in Flask applications
Flask-Login==0.5.0
